Using ACL in font-end

+++ app/Http/Controllers/API/UserController.php
@@ -132,6 +132,8 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize("isAdmin");
+
         $user = User::findOrFail($id);
 
         $user->delete();
